Refactor editor parts into configEditor
